Same

+++ tasks/word2vec/word2vec.cpp
@@ -18,7 +18,7 @@ std::vector<std::string> FindClosestWords(const std::vector<std::string>& words,
     std::vector<std::string> closest_words = {words[1]};
     int64_t current_max_dot_product = DotProduct(vectors[0], vectors[1]);
     for (size_t i = 2; i < vectors.size(); ++i) {
-        int current_dot_product = DotProduct(vectors[0], vectors[i]);
+        int64_t current_dot_product = DotProduct(vectors[0], vectors[i]);
         if (current_dot_product > current_max_dot_product) {
             closest_words.clear();
             closest_words.emplace_back(words[i]);
